Major patch
General changes
 - Reimplemented tracing to use memory mapped files. (It’s fast enough now to keep it enabled all the time. This allows for catching bugs occurring for the first time.)
 - Extended and refactored metrics support (added network I/O and timeouts)
 - Enable full core dumps for OSX, Windows and Linux (see https://audiogridder.com/bug-reports/#submit on how to submit core dumps, as they get large.)
 - Improved concurrency handling and interaction with the main message thread
 - Moved config and log file locations to more appropriate locations on each OS (resolves #175)
 - Auto cleanup of log files and core files
 - Allow to deactivate logging (resolves #175)
 - Fixed offset bug in message send method
 - Improved signal handling

Plugin changes
 - Added plugin statistics
 - Allow remote server restarts
 - Improved remote plugin loading code
 - Avoid any blocking network I/O
 - Relaxed CPU load updates to every 5 seconds and try to fetch them via mDNS rather than sending queries
 - Added a separate MIDI FX plugin to solve some cases where either the instrument or the MIDI plugin was not detected (fixes #177)
 - Fixed bottom cut-off of plugin UIs (fixes #189)
 - Show error reason for plugin load failures
 - Fixed some plugin stack UI and state bugs
 - Fixed mDNS bug that lead to not detecting available servers under certain circumstances

Server changes
 - Added server statistics
 - Allow server restarts
 - Aggregate CPU load over a few seconds
 - Updated plugin ID handling: reinstalled, moved or updated plugins should keep their ID’s which will make restoring DAW sessions more reliable
 - Hide mono plugins from stereo channels
 - Improved I/O layout matching at plugin initialization: waves stereo plugins load on mono channels now
 - Fixed recents list handling
 - Allow parallel plugin loading to improve session load times
 - Skip the first few captured frames to avoid flickering when switching plugin UIs
 - Some improvements in the shutdown behaviour

+++ Plugin/Source/GenericEditor.cpp
@@ -61,7 +61,7 @@ void GenericEditor::resized() {
             c->onChange = [this, i] {
                 auto& locParam = getParameter(i);
                 auto* locComp = dynamic_cast<ComboBox*>(getComponent(i));
-                locParam.setValue(locComp->getSelectedItemIndex());
+                locParam.setValue((float)locComp->getSelectedItemIndex());
                 updateParameter(i);
             };
